pybricks/ble: send() blocking until advertising
This change is made to prevent overflowing the BTchip in a tight loop, i.e., fast consecutive broadcast.send()
The methods broadcast.send() and broadcast.send_bytes() are blocking the python program until the BT chip completed setting the advertising data and starting advertising.
Stopping the advertising after 1 second is not blocking the python program. (In a tight loop this advertising stop is never used).