pbio/broadcast: Implement broadcast logic in pbio.
This reworks the broadcast implementation by moving the high level
logic to pbio, leaving only driver-specific functions in pbdrv.

The previous implementation would crash the hub quite often. Both for
receiving and transmitting, some data could be out of scope before
it was read, or data could be written to areas that had been garbage
collected.

This commit pre-allocates a fixed number of signals statically to
avoid such problems.

This also adds time stamps to incoming messages, which makes the signal
index values more reliable because we know when a message may be
outdated. It also prevents restarting the transmission if the same
message is repeatedly sent in a tight loop.